refactor(infrastructure): move database setup and schema to infrastructure layer
